fixed memory leak from instruction nodes not having virtual destructors

+++ src/ir/node_print.cpp
@@ -10,7 +10,7 @@ void ir::block::block_instruction::print(std::ostream &ostream) const {
 
     inst->print(ostream);
 
-    for (int i = 0; i < operands.size(); i++) {
+    for (size_t i = 0; i < operands.size(); i++) {
         ostream << " ";
         ostream << operands[i];
